+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.18.1 (2026-09-02)
 
 - Load Content-Length tool dependencies without Composer dev autoloading